增加摸鱼检测
